// modules/transfer/modules/transfer-own/enum/own-transfer.enum.ts
export enum EOwnTransferScheduleService {
  OWN = 'transfer2'
}

export enum EOwnTransferViewMode {
  DEFAULT = 'default',
  TRANSACTION_HISTORY = 'transactionHistory',
  SIGNATURE_TRACKING = 'signatureTracking',
  SIGNATURE_TRACKING_MODIFY = 'signatureTrackingModify',
  SIGNATURE_TRACKING_OPERATION = 'signatureTrackingOperation'
}